+++ front/pages/settings/transactions/fields.vue
@@ -68,7 +68,7 @@ const init = () => {
 const toolbar = useToolbar()
 toolbar.init({
   title: 'Transaction fields order',
-  backRoute: RouteConstants.ROUTE_SETTINGS_USER_PREFERENCES_TRANSACTIONS,
+  backRoute: RouteConstants.ROUTE_SETTINGS_TRANSACTION,
 })
 
 onMounted(() => {
