+++ front/pages/settings/ui.vue
@@ -4,25 +4,26 @@
 
     <van-form @submit="onSave" class="">
       <van-cell-group inset>
-        <app-boolean label="Dark theme:" v-model="darkTheme">
+        <div class="van-cell-group-title mb-0">Theme:</div>
+
+        <app-boolean :label="themeText" v-model="darkTheme">
           <template #icon="{ value }">
             <app-icon :size="23" :stroke-width="2.0" :icon="value ? TablerIconConstants.darkTheme : TablerIconConstants.whiteTheme" />
           </template>
         </app-boolean>
-
       </van-cell-group>
 
       <van-cell-group inset>
-        <div class="van-cell-group-title mb-0">List view:</div>
+        <div class="van-cell-group-title mb-0">Transaction list:</div>
 
         <app-select
-            popupTitle="Select what Hero Icons to show"
-            v-model="heroIcons"
-            v-model:showDropdown="isHeroIconsDropdownVisible"
-            :list="heroIconsList"
-            :is-multi-select="true"
-            :columns="1"
-            :has-search="false"
+          popupTitle="Select what Hero Icons to show"
+          v-model="heroIcons"
+          v-model:showDropdown="isHeroIconsDropdownVisible"
+          :list="heroIconsList"
+          :is-multi-select="true"
+          :columns="1"
+          :has-search="false"
         >
           <template #label>
             <div class="flex-center-vertical">
@@ -45,37 +46,26 @@ import { useDataStore } from '~/stores/dataStore'
 import UIUtils from '~/utils/UIUtils'
 import { useToolbar } from '~/composables/useToolbar'
 import RouteConstants from '~/constants/RouteConstants'
-import { NUMBER_FORMAT } from '~/utils/MathUtils.js'
 import TablerIconConstants from '~/constants/TablerIconConstants.js'
 import { saveSettingsToStore, watchSettingsStore } from '~/utils/SettingUtils.js'
+import { HERO_ICONS_LIST } from '~/constants/TransactionConstants.js'
 
 const profileStore = useProfileStore()
 const dataStore = useDataStore()
 
-const numberFormatList = [NUMBER_FORMAT.eu, NUMBER_FORMAT.international]
-const isDropdownNumberFormatVisible = ref(false)
-
+const themeText = computed(() => (darkTheme.value ? 'Dark' : 'Light'))
 const darkTheme = ref(false)
-const numberFormat = ref(null)
-const stripAccents = ref(false)
-const lowerCaseTransactionDescription = ref(false)
-const lowerCaseAccountName = ref(false)
-const lowerCaseCategoryName = ref(false)
-const lowerCaseTagName = ref(false)
+
+const heroIconsList = HERO_ICONS_LIST
+const isHeroIconsDropdownVisible = ref(false)
+const heroIcons = ref([])
 
 const syncedSettings = [
   { store: profileStore, path: 'darkTheme', ref: darkTheme },
-  { store: profileStore, path: 'numberFormat', ref: numberFormat },
-  { store: profileStore, path: 'stripAccents', ref: stripAccents },
-  { store: profileStore, path: 'lowerCaseTransactionDescription', ref: lowerCaseTransactionDescription },
-  { store: profileStore, path: 'lowerCaseAccountName', ref: lowerCaseAccountName },
-  { store: profileStore, path: 'lowerCaseCategoryName', ref: lowerCaseCategoryName },
-  { store: profileStore, path: 'lowerCaseTagName', ref: lowerCaseTagName },
+  { store: profileStore, path: 'heroIcons', ref: heroIcons },
 ]
 
 watchSettingsStore(syncedSettings)
-
-
 
 const onSave = async () => {
   saveSettingsToStore(syncedSettings)
@@ -85,7 +75,7 @@ const onSave = async () => {
 
 const toolbar = useToolbar()
 toolbar.init({
-  title: 'UI preferences',
+  title: 'UI settings',
   backRoute: RouteConstants.ROUTE_SETTINGS,
 })
