--- front/pages/settings/user-preferences-quick-transaction-amounts.vue
@@ -1,66 +1,0 @@
-<template>
-  <div class="app-form">
-    <app-top-toolbar />
-
-    <van-form @submit="onSave" class="">
-      <van-cell-group inset>
-        <div class="van-cell-fake flex-column van-cell">
-          <app-repeater v-model="quickAmountValues" :empty-item="{ value: '' }">
-            <template #content="{ element, index }">
-              <app-field placeholder="Value" v-model="element.value" inputmode="decimal" />
-            </template>
-          </app-repeater>
-        </div>
-      </van-cell-group>
-
-      <app-button-form-save />
-    </van-form>
-  </div>
-</template>
-
-<script setup>
-import { onMounted, ref } from 'vue'
-import { useProfileStore } from '~/stores/profileStore'
-import { useDataStore } from '~/stores/dataStore'
-import UIUtils from '~/utils/UIUtils'
-import { useToolbar } from '~/composables/useToolbar'
-import RouteConstants from '~/constants/RouteConstants'
-
-const profileStore = useProfileStore()
-const dataStore = useDataStore()
-
-const quickAmountValues = ref([])
-
-onMounted(() => {
-  init()
-})
-
-const onSave = async () => {
-  profileStore.quickValueButtons = quickAmountValues.value.map((item) => {
-    let value = sanitizeAmount(item.value)
-    let startsWithOperator = ['-', '+'].includes(value[0])
-    return startsWithOperator ? value : `+${value}`
-  })
-
-  await profileStore.writeProfile()
-
-  UIUtils.showToastSuccess('User preferences saved')
-  init()
-}
-
-const init = () => {
-  quickAmountValues.value = profileStore.quickValueButtons.map((item) => {
-    return { value: item }
-  })
-}
-
-const toolbar = useToolbar()
-toolbar.init({
-  title: 'Quick transaction amounts',
-  backRoute: RouteConstants.ROUTE_SETTINGS_USER_PREFERENCES_TRANSACTIONS,
-})
-
-onMounted(() => {
-  animateSettings()
-})
-</script>
